drm_hwcomposer: provide a common gralloc handle definition
EGL, gralloc, and HWC must all have a common definition of fd's and int's
in native_handle_t to share the fd and width, height, format, etc. of a
dmabuf.

Move the definition into HWC so we aren't dependent on a specific gralloc
implementation and so we don't have to create an importer just for
different native_handle_t layouts. This will allow supporting multiple
gralloc implementations that conform to this layout.

For now, this is aligned with gbm_gralloc's struct. Once we change
gbm_gralloc and mesa to point to this copy, we can make modifications to
the struct.

Signed-off-by: Rob Herring <robh@kernel.org>